+++ tests/e2e/app-files/say_hello.R
@@ -1,3 +1,4 @@
+#' @export
 say_hello <- function(name) {
   paste0("Hello, ", name, "!")
 }