+++ NEWS.md
@@ -1,6 +1,11 @@
 # rhino (development version)
 
 1. Added `sass: custom` configuration option for cleaner integration with `bslib`.
+2. Introduce `format_js()` and `format_sass()` powered by [prettier](https://prettier.io).
+    * **Migration**: Remove the `.rhino` directory if updating Rhino in an existing project.
+    * **Note:** `lint_js()` and `lint_sass()` report styling errors.
+      They _might_ complain about formatting done with `format_js()` and `format_sass()` functions; however, we haven't spotted any issues so far.
+      If you face any problems with this, then please [raise an issue on GitHub](https://github.com/Appsilon/rhino/issues/new/choose)
 
 # [rhino 1.8.0](https://github.com/Appsilon/rhino/releases/tag/v1.8.0)